Major Refactoring of the Data structure
Fixes #100 #87

Changed the Datatype of label to "Label" from which several subclasses
- TextLabel
- TextListLabel
- DateTimeLabel

Changed the Datatype of Datasets.data to List<? extends ValueType>.
ValueType has several subclasses
- StringValue
- IntegerValue
- FloatValue
- DoubleValue
- PointValue
- BubbleValue

As a result many methods are adjusted to support the new strucutre. Some
tests are provided for this strucuture.

Added Time scale - Line example.